# apidocs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8000","description":"Generated server url"}],"paths":{"/api/users/update/{id}":{"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/users/create":{"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/identity/introspect":{"post":{"tags":["authentication-controller"],"operationId":"introspect","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntrospectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseIntrospectResponse"}}}}}}},"/api/identity/authenticate":{"post":{"tags":["authentication-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthenticationResponse"}}}}}}},"/api/users/profile":{"get":{"tags":["user-controller"],"operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/users/getDetail/{id}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}}}}}},"/api/users/getAll":{"get":{"tags":["user-controller"],"operationId":"getAllUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUserResponse"}}}}}}},"/api/users/delete/{id}":{"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UserUpdateRequest":{"type":"object","properties":{"password":{"type":"string","maxLength":2147483647,"minLength":5},"fullName":{"type":"string"},"dob":{"type":"string","format":"date"}}},"ApiResponseUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"$ref":"#/components/schemas/UserResponse"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"fullName":{"type":"string"},"dob":{"type":"string","format":"date"},"roles":{"type":"array","items":{"type":"string"},"uniqueItems":true}}},"UserCreationRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":50,"minLength":3,"pattern":"^[a-zA-Z][a-zA-Z0-9._-]{2,}$"},"password":{"type":"string","maxLength":2147483647,"minLength":5},"fullName":{"type":"string"},"dob":{"type":"string","format":"date"}},"required":["password","username"]},"IntrospectRequest":{"type":"object","properties":{"token":{"type":"string","minLength":1}},"required":["token"]},"ApiResponseIntrospectResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"$ref":"#/components/schemas/IntrospectResponse"}}},"IntrospectResponse":{"type":"object","properties":{"valid":{"type":"boolean"}}},"AuthenticationRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"ApiResponseAuthenticationResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"$ref":"#/components/schemas/AuthenticationResponse"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"},"authenticated":{"type":"boolean"}}},"ApiResponseListUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}